Домашнее задание 22 спринта

+++ app/src/main/java/com/practicum/playlistmaker/library/data/FavoritesApiImpl.kt
@@ -1,22 +1,31 @@
 package com.practicum.playlistmaker.library.data
 
 import com.practicum.playlistmaker.db.data.AppDatabase
+import com.practicum.playlistmaker.db.data.FavoriteTrack
 import com.practicum.playlistmaker.library.domain.FavoritesApi
 import com.practicum.playlistmaker.search.domain.entities.Track
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
 
-class FavoritesApiImpl(private val appDatabase: AppDatabase) : FavoritesApi {
+class
+
+
+FavoritesApiImpl(private val appDatabase: AppDatabase) : FavoritesApi {
 
     override suspend fun addNewFavorite(track: Track) {
         appDatabase.getTrackDao().insertNewTrack(track)
+        val favorite = FavoriteTrack(track.trackId)
+        appDatabase.getFavoriteDao().addToFavorites(favorite)
     }
 
     override suspend fun deleteFavorite(trackId: Long) {
-        appDatabase.getTrackDao().deleteTrackById(trackId)
+        appDatabase.getFavoriteDao().deleteTrackById(trackId)
+        if (!appDatabase.getPlaylistTrackCrossRefDao().isTrackInAnyPlaylist(trackId)) {
+            appDatabase.getTrackDao().deleteTrackById(trackId)
+        }
     }
 
     override suspend fun getAllFavorites(): Flow<List<Track>> = flow {
-        emit( appDatabase.getTrackDao().getTracks())
+        emit( appDatabase.getFavoriteDao().getAllFavoritesOrdered())
     }
 }
